src/main/java/com/fabrick/bankapp/dto/transferDto/Address.java
package com.fabrick.bankapp.dto.transferDto;

import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonProperty;

import java.io.Serializable;

public class Address implements Serializable {
    private String address;
    private String city;
    private String countryCode;

    @JsonCreator
    public Address(@JsonProperty("address") String address,
                   @JsonProperty("city") String city,
                   @JsonProperty("countryCode") String countryCode) {
        this.address = address;
        this.city = city;
        this.countryCode = countryCode;
    }

    public String getAddress() {
        return address;
    }

    public void setAddress(String address) {
        this.address = address;
    }

    public String getCity() {
        return city;
    }

    public void setCity(String city) {
        this.city = city;
    }

    public String getCountryCode() {
        return countryCode;
    }

    public void setCountryCode(String countryCode) {
        this.countryCode = countryCode;
    }


}

